Refactor: move methods to buff.cpp
