+++ variables.hpp
@@ -689,7 +689,6 @@ ELONA_EXTERN(std::string userfile);
 ELONA_EXTERN(std::string usermsg);
 int access_character_info();
 bool actionsp(int, int);
-int addbuff(int = 0, int = 0, int = 0, int = 0);
 int breath_list();
 int calcincome(int = 0);
 int calcmagiccontrol(int = 0, int = 0);
@@ -717,7 +716,6 @@ void damage_insanity(int, int);
 void dmgsp(int, int);
 int efstatusfix(int = 0, int = 0, int = 0, int = 0);
 int eleinfo(int = 0, int = 0);
-int findbuff(int = 0, int = 0);
 int findunid(const std::string&);
 int fltsetdungeon();
 int gameover();
@@ -912,7 +910,6 @@ void cslineup();
 void cursor_check();
 void cutname(std::string&, int = 0);
 void delbottomcard(int = 0);
-void delbuff(int = 0, int = 0);
 void delmef(int = 0);
 void dipcursed(int = 0, int = 0);
 void dmgcard(int = 0, int = 0);
